Abstract release logic away from the get_releases route
