+++ orlo/queries.py
@@ -1,13 +1,123 @@
 from __future__ import print_function
+import datetime
+import arrow
 from orlo import app
 from orlo.orm import db, Release, Platform, Package, release_platform
-from orlo.exceptions import OrloError
+from orlo.exceptions import OrloError, InvalidUsage
 
 __author__ = 'alforbes'
 
 """
 Functions in this file are about generating summaries of data
 """
+
+
+def apply_filters(query, args):
+    """
+    Apply filters to a query
+
+    :param query: Query object to apply filters to
+    :param args: Dictionary of arguments, usually request.args
+
+    :return: filtered query object
+    """
+
+    for field, value in args.iteritems():
+        if field == 'latest':  # this is not a comparison
+            continue
+
+        if field.startswith('package_'):
+            # Package attribute. Ensure source query does a join on Package.
+            db_table = Package
+            field = '_'.join(field.split('_')[1:])
+        else:
+            db_table = Release
+
+        comparison = '=='
+        time_absolute = False
+        time_delta = False
+        strip_last = False
+        sub_field = None
+
+        if field.endswith('_gt'):
+            strip_last = True
+            comparison = '>'
+        if field.endswith('_lt'):
+            strip_last = True
+            comparison = '<'
+        if field.endswith('_before'):
+            strip_last = True
+            comparison = '<'
+            time_absolute = True
+        if field.endswith('_after'):
+            strip_last = True
+            comparison = '>'
+            time_absolute = True
+        if 'duration' in field.split('_'):
+            time_delta = True
+        if field == 'platform':
+            field = 'platforms'
+            comparison = 'any'
+            sub_field = Platform.name
+
+        if strip_last:
+            # Strip anything after the last underscore inclusive
+            field = '_'.join(field.split('_')[:-1])
+
+        filter_field = getattr(db_table, field)
+
+        # Booleans
+        if value in ('True', 'true'):
+            value = True
+        if value in ('False', 'false'):
+            value = False
+
+        # Time related
+        if time_delta:
+            value = datetime.timedelta(seconds=int(value))
+        if time_absolute:
+            value = arrow.get(value)
+
+        # Do comparisons
+        app.logger.debug("Filtering: {} {} {}".format(filter_field, comparison, value))
+        if comparison == '==':
+            query = query.filter(filter_field == value)
+        if comparison == '<':
+            query = query.filter(filter_field < value)
+        if comparison == '>':
+            query = query.filter(filter_field > value)
+        if comparison == 'any':
+            query = query.filter(filter_field.any(sub_field == value))
+
+    return query
+
+
+def releases(**kwargs):
+    """
+    Return whole releases, based on filters
+
+    :param kwargs: Request arguments
+    :return:
+    """
+
+    if any(field.startswith('package_') for field in kwargs.keys()):
+        query = db.session.query(Release).join(Package)
+    else:
+        # No need to join on package if none of our params need it
+        query = db.session.query(Release)
+
+    try:
+        query = apply_filters(query, kwargs)
+    except AttributeError as e:
+        raise InvalidUsage("An invalid field was specified: {}".format(e.message))
+
+    if kwargs.get('latest', False):
+        # sort descending so first is most recent
+        query = query.order_by(Release.stime.desc()).limit(1)
+    else:  # ascending
+        query = query.order_by(Release.stime.asc())
+
+    return query
 
 
 def user_summary(platform=None):
@@ -334,3 +444,5 @@ def platform_list():
         .join(release_platform)
 
     return query
+
+
